Projeto 02 - Ingestão de Dados de API/pyproject.toml
[tool.poetry]
name = "function-to-ingest-data"
version = "0.1.0"
description = ""
authors = ["iuryrosal <iuryrosal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-secret-manager = "^2.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
